requirements.txt
# Core
pandas==2.3.3
numpy==2.0.2
faker==37.12.0

# Kafka
kafka-python==2.0.2

# PySpark
pyspark==3.5.1

# ML
scikit-learn==1.5.0
xgboost==2.0.3
joblib==1.4.2

# API
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1

# Airflow
apache-airflow==2.9.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.30

# Testing
pytest==8.2.0
pytest-html==4.1.1
requests==2.32.2
selenium==4.21.0

# Utilities
python-dotenv==1.0.1
great-expectations==0.18.15
hashlib2==1.0.2
